Line Trace for Widget visibility (1/2)
The Weapon widget is now displaying above the weapon when the player aim at it (crosshair location).
For now, as a proof of concept, the line tracing is performed every frame... I know. We'll have to refactor this later. Also, we need to hide the widget when not aiming the weapon anymore
Refactor time incoming <3